% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructSExprFromDEE2Data.R
\name{constructSExprFromDEE2Data}
\alias{constructSExprFromDEE2Data}
\title{Construct the SummarizedExperiment object for the raw counts data.}
\usage{
constructSExprFromDEE2Data(
  sampleInfoPath,
  dee2DataDir,
  species = c("athaliana", "celegans", "dmelanogaster", "drerio", "ecoli", "hsapiens",
    "mmusculus", "rnorvegicus", "scerevisiae"),
  overwrite = FALSE
)
}
\arguments{
\item{sampleInfoPath}{a path to a .ods file that stores category, 
general_cell_type, parental_cell_type and SRR_accession for each 
sample}

\item{dee2DataDir}{directory to store files storing DEE2 data}

\item{species}{the species of the DEE2 data}

\item{overwrite}{a boolean parameter. If it is true, 'constructSExprFromDEE2Data'
function will overwrite the files in the directory specified by parameter
dee2DataDir. By default, the value is false.}

\item{sExpr}{SummarizedExperiment object that is created via DEE2 R interface 
or restored from downloaded files.}

\item{counts.cutoff}{threshold of filtering the}
}
\description{
Construct the SummarizedExperiment object for the raw counts data.
}
