% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2Normalization.R
\name{Deseq2Normalization}
\alias{Deseq2Normalization}
\title{Deseq2 normalization for the selected DEE2 data
'In some experiment, there might be gene-dependent dependencies which vary 
across samples. For instance, GC-content bias or length bias might vary across 
samples coming from different labs or processed at different times...'}
\usage{
Deseq2Normalization(dee2Data)
}
\description{
Deseq2 normalization for the selected DEE2 data
'In some experiment, there might be gene-dependent dependencies which vary 
across samples. For instance, GC-content bias or length bias might vary across 
samples coming from different labs or processed at different times...'
}
\references{
http://www.sthda.com/english/wiki/rna-sequencing-data-analysis-counting-normalization-and-differential-expression#normalization-using-deseq2-size-factors

https://uclouvain-cbio.github.io/BSS2019/rnaseq_gene_summerschool_belgium_2019.html
}
