% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2NormalizationFromTranscriptLevelData.R
\name{DESeq2NormalizationFromTranscriptLevelData}
\alias{DESeq2NormalizationFromTranscriptLevelData}
\title{DESeq2 normalization based on gene-level count data summarized from 
transcript-level estimates}
\usage{
DESeq2NormalizationFromTranscriptLevelData(dee2Data, counts.cutoff = 10)
}
\arguments{
\item{dee2Data}{SummarizedExperiment}
}
\description{
DESeq2 normalization based on gene-level count data summarized from 
transcript-level estimates
}
\references{
https://bioconductor.org/packages/release/bioc/html/tximport.html

http://bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html

https://uclouvain-cbio.github.io/BSS2019/rnaseq_gene_summerschool_belgium_2019.html
}
