% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDEE2Data.R
\name{.constructSExprFromDEE2Data}
\alias{.constructSExprFromDEE2Data}
\title{.constructSExprFromDEE2Data function}
\usage{
.constructSExprFromDEE2Data(dee2Data, samplesInfoDFrame)
}
\arguments{
\item{dee2Data}{a vector of dataframes that stores DEE2 data.}
}
\value{
dee2Data
}
\description{
.constructSExprFromDEE2Data function
}
