% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.downloadAbundanceTsv}
\alias{.downloadAbundanceTsv}
\title{.downloadAbundanceTsv function}
\usage{
.downloadAbundanceTsv(species, srrAccessions, outDir = NULL)
}
\arguments{
\item{outDir}{}
}
\value{
abundancefiles
}
\description{
.downloadAbundanceTsv function
}
