% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDEE2Data.R
\name{.excludeFailedSamples}
\alias{.excludeFailedSamples}
\title{.excludeFailedSamples function}
\usage{
.excludeFailedSamples(dee2Data)
}
\arguments{
\item{dee2Data}{a vector of dataframes that stores DEE2 data.}
}
\value{
dee2Data
}
\description{
.excludeFailedSamples function
}
