% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.filterData}
\alias{.filterData}
\title{Calculates z-score. We use z-score as a measure of how close the standards
cluster to each other}
\usage{
.filterData(sExpr, counts.cutoff = 5, excludeFail = TRUE)
}
\arguments{
\item{sExpr}{raw sExpr data object}

\item{counts.cutoff}{the threshold for filtering genes. For a specific gene,
the total cound}
}
\value{
sExpr
}
\description{
Calculates z-score. We use z-score as a measure of how close the standards
cluster to each other
}
