% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2Data.R
\name{.orderSamples}
\alias{.orderSamples}
\title{order all dataframes in 'dee2Data' by the same order of samples in 
'GeneCount' dataframe}
\usage{
.orderSamples(dee2Data)
}
\arguments{
\item{dee2Data}{a vector of dataframes that stores DEE2 data.}
}
\value{
Updated vector
}
\description{
order all dataframes in 'dee2Data' by the same order of samples in 
'GeneCount' dataframe
}
