% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadDEE2Data.R
\name{DownloadDEE2Data}
\alias{DownloadDEE2Data}
\title{Download DEE2 Data.}
\usage{
DownloadDEE2Data(species, sRRAccessions, bulkDownloading = FALSE)
}
\description{
By default, 'bulkDownloading' is set to false, as the current version of DEE2 
could not get consistent data with bulk downloading.
}
\examples{
## Load the dependencies
library(DEE2HsapienData)

## Create a list of SRR Accessions of which we want to retrieve the counts 
from DEE2 library
SRRvec = c("SRR1783836", "SRR1783837", "SRR1783838", "SRR1999221", "SRR2153338", 
           "SRR2153409", "SRR2153289")

## Creat a 'temp' directory in current working directory if it is not created
## already. In this tutorial, we use it as the workspace 
dee2Data <- DownloadDEE2Data('hsapiens', SRRvec)
}
\references{
https://github.com/markziemann/dee2
}
