% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptLevelAnalysis.R
\name{edgeRTranscriptAnalysis}
\alias{edgeRTranscriptAnalysis}
\title{edgeR normalization based on Transcript-level count data}
\usage{
edgeRTranscriptAnalysis(species, srrAccessions, txInfo, outDir)
}
\description{
edgeR normalization based on Transcript-level count data
}
\references{
https://bioconductor.org/packages/release/bioc/html/tximport.html
}
