% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2Data.R
\name{getDEE2Data}
\alias{getDEE2Data}
\title{Construct a 'list' object that includes seven 'data.frame' objects 
( "GeneCounts", "TxCounts", "GeneInfo", "TxInfo", "QcMx", "MetadataSummary" 
and "MetadataFull"), and a 'SummarizedExperiment' object 'sExpr'. In the case 
when there are SRR accessions not present in DEE2 database or not matching
the 'species' parameter, an list object 
named 'absent' will be in the returned list containing those accessions. The 
'sExpr' object in the returned list contains an assay containing the same 
data as in "GeneCounts" object. The column data of 'sExpr' is a 'data.frame' 
that combines the data in the spreadsheet indicated by 'sampleInfoPath' and 
in 'MetadataFull' object. The row data of 'sExpr' includes the same data as 
in 'GeneInfo' object.}
\usage{
getDEE2Data(
  sampleInfoPath,
  outFile,
  species = c("athaliana", "celegans", "dmelanogaster", "drerio", "ecoli", "hsapiens",
    "mmusculus", "rnorvegicus", "scerevisiae")
)
}
\arguments{
\item{sampleInfoPath}{A path to a spreadsheet file that stores 'category', 
'general_cell_type', 'parental_cell_type' and 'SRR_accession' for each 
sample.}

\item{outFile}{By default, this parameter is NULL. If it is defined, then 
files that contains DEE2 data will be downloaded to the current working 
directory in a 'zip' file named as the value of this parameter.}

\item{species}{the species of the DEE2 data. Available species includes 
'athaliana', 'celegans','dmelanogaster','drerio','ecoli', 'hsapiens',
'mmusculus','rnorvegicus' and 'scerevisiae'.}
}
\description{
for details of  "GeneCounts", "TxCounts", "GeneInfo", "TxInfo", "QcMx", 
"MetadataSummary" and "MetadataFull" dataframs, refer to 
https://github.com/markziemann/dee2/blob/master/AccessDEEfromR.md
}
\examples{
## Load the dependencies
library(SummarizedExperiment)

## Creat a 'temp' directory in current working directory if it is not created
## already. In this tutorial, we use it as the workspace 
tempDirPath <- paste(getwd(), 'temp', sep = '/')
dir.create(tempDirPath)

## get the 'list' object. 
dee2Data <- getDEE2Data(
 system.file("extdata", "SelectedDEE2_v2.ods", package = "DEE2HsapienData"),
 paste(tempDirPath, 'DEE2Data', sep = '/') , 
 'hsapiens')
}
\references{
https://github.com/markziemann/dee2/blob/master/AccessDEEfromR.md
}
