% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizedCountsHistogramPlot.R
\name{NormalizedCountsHistogramPlot}
\alias{NormalizedCountsHistogramPlot}
\title{Histogram Plots of Normalized Counts}
\usage{
NormalizedCountsHistogramPlot(sExpr)
}
\arguments{
\item{sExpr}{SummarizedExperiment object that is created via DEE2 R interface}
}
\description{
Histogram Plots of Normalized Counts
}
