% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PcaPlot.R
\name{PcaPlot}
\alias{PcaPlot}
\title{PCA analysis of the count data and plot the result in 2D / 3D scatter graphs}
\usage{
PcaPlot(sExpr, plot = "3d")
}
\arguments{
\item{sExpr}{a SummarizedExperiment object that contains count data along 
with gene information and sample metadata.}

\item{plot}{the option of the plot of the PCA analysis result, which should
be either '2d' or '3d'}
}
\description{
PCA analysis of the count data and plot the result in 2D / 3D scatter graphs
}
