% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetGeneMetadata.R
\name{SetGeneMetadata}
\alias{SetGeneMetadata}
\title{SetGeneMetadata}
\usage{
SetGeneMetadata(sExpr, species)
}
\arguments{
\item{sExpr}{SummarizedExperiment}

\item{species}{character}
}
\description{
SetGeneMetadata
}
