% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMMNormalizationFromTranscriptLevelData.R
\name{TMMNormalizationFromTranscriptLevelData}
\alias{TMMNormalizationFromTranscriptLevelData}
\title{TMM normalization based on gene-level count data summarized from 
transcript-level estimates}
\usage{
TMMNormalizationFromTranscriptLevelData(dee2Data)
}
\description{
TMM normalization based on gene-level count data summarized from 
transcript-level estimates
}
\references{
https://bioconductor.org/packages/release/bioc/html/tximport.html
}
