% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsnePlot.R
\name{TsnePlot}
\alias{TsnePlot}
\title{tsne analysis of the count data}
\usage{
TsnePlot(sExpr, plot = "3d")
}
\arguments{
\item{sExpr}{a SummarizedExperiment object that contains count data along 
with gene information and sample metadata.}

\item{plot}{the option of the plot of the PCA analysis result, which should
be either '2d' or '3d'}
}
\description{
tsne analysis of the count data
}
