% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tximportDee2Data.R
\name{tximportDee2Data}
\alias{tximportDee2Data}
\title{Download abundance.tsv files (transcript-level estimate) for specified species 
and SSR Accessions and get the gene-level estimate}
\usage{
tximportDee2Data(species, dee2Data, outDir = NULL)
}
\arguments{
\item{txInfo}{dee2Data$TxInfo}
}
\value{
dee2Data
}
\description{
Download abundance.tsv files (transcript-level estimate) for specified species 
and SSR Accessions and get the gene-level estimate
}
\examples{

## This operation may requires up to 8GB of memory
txi <- tximportDee2Data('hsapiens', dee2Data, )
}
\references{
https://bioconductor.org/packages/devel/bioc/vignettes/tximport/inst/doc/tximport.html
}
