% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructSExprFromDEE2Data.R
\name{constructSExprFromDEE2Data}
\alias{constructSExprFromDEE2Data}
\title{Construct the SummarizedExperiment object for the raw counts data.}
\usage{
constructSExprFromDEE2Data(
  sampleInfoPath,
  outFile,
  species = c("athaliana", "celegans", "dmelanogaster", "drerio", "ecoli", "hsapiens",
    "mmusculus", "rnorvegicus", "scerevisiae")
)
}
\arguments{
\item{sampleInfoPath}{A path to a spreadsheet file that stores 'category', 
'general_cell_type', 'parental_cell_type' and 'SRR_accession' for each 
sample.}

\item{outFile}{By default, this parameter is NULL. If it is defined, then 
files that contains DEE2 data will be downloaded to the current working 
directory in a 'zip' file named as the value of this parameter.}

\item{species}{the species of the DEE2 data. Available species includes 
'athaliana', 'celegans','dmelanogaster','drerio','ecoli', 'hsapiens',
'mmusculus','rnorvegicus' and 'scerevisiae'.}
}
\description{
Construct the SummarizedExperiment object for the raw counts data.
}
\references{
https://github.com/markziemann/dee2/blob/master/AccessDEEfromR.md
}
