% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2Normalization.R
\name{DESeq2Normalization}
\alias{DESeq2Normalization}
\title{DESeq2 normalization for the selected DEE2 data}
\usage{
DESeq2Normalization(sExpr, counts.cutoff = 5, excludeFail = TRUE)
}
\arguments{
\item{sExpr}{SummarizedExperiment object that is created via DEE2 R interface 
or restored from downloaded files.}

\item{counts.cutoff}{threshold of filtering the}
}
\description{
DESeq2 normalization for the selected DEE2 data
}
