% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.filterData}
\alias{.filterData}
\title{.filterData function}
\usage{
.filterData(sExpr, counts.cutoff = 10)
}
\arguments{
\item{sExpr}{raw sExpr data object}

\item{counts.cutoff}{the threshold for filtering genes. For a specific gene,
the total cound}
}
\value{
sExpr
}
\description{
.filterData function
}
