% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDEE2Data.R
\name{GetDEE2Data}
\alias{GetDEE2Data}
\title{Get a 'SummarizedExperiment' object 'sExpr' from DEE2 library and set its 
gene info based on the data from biomaRt.}
\usage{
GetDEE2Data(
  sampleInfoPath,
  species,
  counts = "GeneCounts",
  bulkDownloading = FALSE
)
}
\arguments{
\item{sampleInfoPath}{A path to a spreadsheet file that stores 'category', 
'general_cell_type', 'parental_cell_type' and 'SRR_accession' for each 
sample.}

\item{species}{the species of the DEE2 data. Available species includes 
'athaliana', 'celegans','dmelanogaster','drerio','ecoli', 'hsapiens',
'mmusculus','rnorvegicus' and 'scerevisiae'.}

\item{counts}{either 'GeneCounts' or 'Tx2Gene'}
}
\description{
Get a 'SummarizedExperiment' object 'sExpr' from DEE2 library and set its 
gene info based on the data from biomaRt.
}
\examples{
## Load the dependencies
library(SummarizedExperiment)

## Creat a 'temp' directory in current working directory if it is not created
## already. In this tutorial, we use it as the workspace 
tempDirPath <- paste(getwd(), 'temp', sep = '/')
dir.create(tempDirPath)

## get the 'list' object. 
dee2Data <- GetDEE2Data(
 system.file("extdata", "SelectedDEE2_v2.ods", package = "DEE2HsapienData"),
 paste(tempDirPath, 'DEE2Data', sep = '/') , 
 'hsapiens')
}
\references{
https://github.com/markziemann/dee2/blob/master/AccessDEEfromR.md
}
