% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizedCountsHistogramPlot.R
\name{NormalizedCountsHistogramPlot}
\alias{NormalizedCountsHistogramPlot}
\title{Histogram Plots of Normalized Counts}
\usage{
NormalizedCountsHistogramPlot(dee2Data, breaks = 100)
}
\arguments{
\item{sExpr}{SummarizedExperiment object that is created via DEE2 R interface}

\item{plot}{the type of plot of the PCA analysis result}
}
\description{
Histogram Plots of Normalized Counts
}
