% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{PCA analysis of normalized count data and plot the result in 2D / 3D scatter
graphs}
\usage{
pcaPlot(sExpr, plot = c("3d", "2d"))
}
\arguments{
\item{sExpr}{a SummarizedExperiment object that contains filtered and normalized
count data along with gene information and sample metadata.}

\item{plot}{the type of plot of the PCA analysis result}
}
\description{
PCA analysis of normalized count data and plot the result in 2D / 3D scatter
graphs
}
