% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMMNormalization.R
\name{TMMNormalization}
\alias{TMMNormalization}
\title{TMM normalization for the selected DEE2 data}
\usage{
TMMNormalization(sExpr)
}
\arguments{
\item{sExpr}{SummarizedExperiment object that is created via DEE2 R interface 
or restored from downloaded files.}

\item{counts.cutoff}{threshold of filtering the}
}
\description{
This function allows you to express your love of cats.
}
\references{
https://www.bioconductor.org/packages/release/bioc/vignettes/edgeR/inst/doc/edgeRUsersGuide.pdf
}
