% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TmmNormalization.R
\name{TmmNormalization}
\alias{TmmNormalization}
\title{TMM normalization for the selected DEE2 data}
\usage{
TmmNormalization(sExpr, counts.cutoff = 5, excludeFail = TRUE)
}
\arguments{
\item{sExpr}{SummarizedExperiment object that is created via DEE2 R interface 
or restored from downloaded files.}

\item{counts.cutoff}{threshold of filtering the}
}
\description{
This function allows you to express your love of cats.
}
