% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TmmNormalizationFromTranscriptLevelData.R
\name{TmmNormalizationFromTranscriptLevelData}
\alias{TmmNormalizationFromTranscriptLevelData}
\title{TMM normalization based on gene-level count data summarized from 
transcript-level estimates}
\usage{
TmmNormalizationFromTranscriptLevelData(dee2Data)
}
\description{
TMM normalization based on gene-level count data summarized from 
transcript-level estimates
}
\references{
https://bioconductor.org/packages/release/bioc/html/tximport.html
}
